refactor[#150]: Enum 원복
