+++ src/main/java/com/plog/plogbackend/global/common/entity/BaseTimeStatusEntity.java
@@ -1,6 +1,6 @@
 package com.plog.plogbackend.global.common.entity;
 
-import com.plog.plogbackend.global.common.enums.EntityStatus;
+import com.plog.plogbackend.global.common.Enum.EntityStatus;
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
 import lombok.Getter;
